+++ src/app/shared/models/localized.type.ts
@@ -1,0 +1,3 @@
+import { Locale } from "../../search-module/models/locale.interface";
+
+export type Localized = Locale;